+++ CHANGELOG.md
@@ -17,5 +17,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix how `start` option was handled from `bs_restore2` instruction: last saved match offset was
   used instead of match starting offset, causing some bytes being skipped.
 - Fix another potential bug when doing pattern matching using code compiled with OTP 21.
+- [ESP32] [UART]: Allow using different pins for rx, tx, cts and rts.
 
 ## [0.5.0] - 2022-03-22